+++ apps/carp_mobile_sensing_app/lib/src/blocs/sensing_bloc.dart
@@ -8,20 +8,17 @@ class SensingBLoC {
   final String uri = "https://cans.cachet.dk:443";
   final String testStudyId = "#18 - activity recognition";
 
-  final Sensing sensing = Sensing();
-  final StudyManager manager = LocalStudyManager();
-
   Study _study;
   Study get study => _study;
 
   /// Is sensing running, i.e. has the study executor been resumed?
-  bool get isRunning => (sensing.controller != null) && sensing.controller.executor.state == ProbeState.resumed;
+  bool get isRunning => (Sensing().controller != null) && Sensing().controller.executor.state == ProbeState.resumed;
 
   /// Get the study for this app.
   StudyModel get studyModel => study != null ? StudyModel(study) : null;
 
   /// Get a list of running probes
-  Iterable<ProbeModel> get runningProbes => sensing.runningProbes.map((probe) => ProbeModel(probe));
+  Iterable<ProbeModel> get runningProbes => Sensing().runningProbes.map((probe) => ProbeModel(probe));
 
   /// Get the data model for this study.
   DataModel get data => null;
@@ -29,15 +26,15 @@ class SensingBLoC {
   void init() async {
     // set global debug level
     globalDebugLevel = DebugLevel.DEBUG;
-    _study ??= await manager.getStudy(testStudyId);
+    _study ??= await Sensing().getStudy(testStudyId);
     debug('Study : $study');
-    await sensing.init();
+    await Sensing().initialize();
   }
 
-  void resume() async => sensing.controller.resume();
-  void pause() => sensing.controller.pause();
-  void stop() async => sensing.stop();
-  void dispose() async => sensing.stop();
+  void resume() async => Sensing().controller.resume();
+  void pause() => Sensing().controller.pause();
+  void stop() async => Sensing().controller.stop();
+  void dispose() async => Sensing().controller.stop();
 }
 
 final bloc = SensingBLoC();