+++ apps/carp_mobile_sensing_app/lib/src/models/study_model.dart
@@ -11,16 +11,16 @@ class StudyModel {
   String get dataEndpoint => study.dataEndPoint.toString();
 
   /// Events on the state of the study executor
-  Stream<ProbeState> get studyExecutorStateEvents => bloc.sensing.controller.executor.stateEvents;
+  Stream<ProbeState> get studyExecutorStateEvents => Sensing().controller.executor.stateEvents;
 
   /// Current state of the study executor (e.g., resumed, paused, ...)
-  ProbeState get studyState => bloc.sensing.controller.executor.state;
+  ProbeState get studyState => Sensing().controller.executor.state;
 
   /// Get all sesing events (i.e. all [Datum] objects being collected).
-  Stream<Datum> get samplingEvents => bloc.sensing.controller.events;
+  Stream<Datum> get samplingEvents => Sensing().controller.events;
 
   /// The total sampling size so far since this study was started.
-  int get samplingSize => bloc.sensing.controller.samplingSize;
+  int get samplingSize => Sensing().controller.samplingSize;
 
   StudyModel(this.study)
       : assert(study != null, 'A StudyModel must be initialized with a real Study.'),